In this example, each HTML file is treated as a separate SCO. The SCOs are aggregated into four items that represent the different topics covered within the course. Some things to notice in this example:
Use of aggregations to form a logical course hierarchy.
Use of the parameters attribute to pass querystring parameters into the shared quiz resource.
The use of resources typed as assets instead of SCOs because they do not communicate with the LMS.
The use of a common shared asset containing files that are reused across many SCOs. This asset is referenced through dependencies on the other resources.
Since this example relies on the LMS for navigation, it is the first functional example that can be launched and run in an LMS. To be a complete and fully-functional course, it still needs some more work. That will come later. For now, try importing it into the SCORM Cloud to see how it looks.