Sequencing Strategy
This example demonstrates a sequencing strategy that requires the learner to visit all SCOs in order. Once a SCO has been visited, the learner can jump backwards to review material, but the learner cannot jump ahead until the prerequisites are met.

Rollup Strategy

Completion: All activities must be completed for the course to be completed.
Satisfaction: All activities must be satisfied for the course to be satisfied. Activities without an assessment are satisfied when they are completed. Activities with an assessment measure satisfaction based on the outcome of the assessment.
Score: No score is rolled up, all activities have an objective measure weight of zero.
Also Notice:

The use of a sequencing collection to define sequencing rules that are common to every activity.
The use of global objectives to track satisfaction of learning objectives (which are also the prerequisites).
The sometimes counter-intuitive use of the “not” operator.
The use of precondition sequencing rules to disable activities until their prerequisites are met.