Sequencing Strategy
This example maintains the sequencing strategy from the Forced Sequential Order example. All SCOs must be visited in order initially, but once a SCO is completed, the learner may jump back to it at any time.

Rollup Strategy

The course can only be completed and satisfied by completing and satisfying the post test. Because of the sequencing strategy that requires all other activities to be completed before the post test can be accessed, by implication, all activities must be completed for the course to be completed. The score for the course is entirely dependent on the score from the post test. This rollup strategy is implemented by setting every activity except for the post test to not contribute to any rollup. When there is only one activity that contributes to rollup in a cluster, then cluster will receive that activity’s status by default during rollup.