Sequencing Strategy
The learner can freely navigate through all activities in the course.
The pretest can only be attempted one time.
Once any test is passed, all tests are disabled.
The post test can not be attempted until all content is completed.
Rollup Strategy

Completion, satisfaction and score are all tied to the most recently completed test.
The learner can “test out” by completing the pretest.
If the learner doesn’t test out, achieving completion requires completing all SCOs. This is done implicitly by putting in sequencing rules that don’t allow another test to be taken until all SCOs are completed.
Also Notice:

The use of an invisible dummy clusters to simplify sequencing and rollup rules.
The use of the attempt limit sequencing attribute in conjunction with a precondition rule to disable the pretest after one attempt. Note how this requires setting cmi.exit to normal instead of suspend to end the attempt.
The use of a global objective shared between the pre and post tests to record satisfaction and score.