SCORM 2004 4th Edition Features This example demonstrates some of the powerful new features available in SCORM 2004 4th Edition:
Sequencing Strategy

This example maintains the sequencing strategy from the Forced Sequential Order and Post Test Rollup examples. All SCOs must be visited in order initially, but once a SCO is completed, the learner may jump back to it at any time.

To demonstrate the new 4th Edition “jump” navigation request, the learner can now skip straight to the test at anytime from within a SCO, bypassing all sequencing rules. If the learner is taking the content though, it still must be taken in order.
To demonstrate the additional data that can be stored on global objectives, this example uses completion status as the driver for prerequisites. In prior editions of SCORM 2004, only success status was available on global objectives. In this case, we simply need the learner to complete the SCO not necessarily satisfy it.
Rollup Strategy

This example maintains the rollup strategy from the Forced Sequential Order and Post Test Rollup examples. It still works even though the sequencing strategy now allows for jump request because the rollup is entirely dependent on the post test.
To demonstrate the new rollup of progress measure, this example now includes weights on each SCO that represent their overall contribution to completion.
Also Notice

To demonstrate the value of the new shared data buckets in 4th Edition, we added a notes feature that allows the learner to take brief notes and have them be available in any SCO.